fix #24: hover hints are red not yellow
This was an easy fix, but investigation led to other issues which also
needed to be fixed.

1. The base problem was that the hover hint color for both "two" and
"many" were wrong: many was yellow (instead of red) and "two" was
green (!) (instead of yellow).  This was easy to fix and was probably a
typo in an earlier commit.

2. While investigating, it became clear the the select hint for "zero"
had been omitted when the possibility of 0-value squares was introduced.
So a red select hint for 0 was introduced.

3. Then an examination of the javascript code indicated that we never
hint a square as having two values, even when there are no single-value,
unassigned squares.  This was a bug.  So code was introduced to count
the number of single-value, unassigned squares in the puzzle, and to
update that count after an assignment.