fix #32: Assignment & Puzzle Construction don't agree about solvability
We had a missing check in group analyze: we weren't making sure that the bindings of the analyzed group were consistent with those of other groups.  Fixing this required generalizing the error message associate with the DuplicateGroupValuesCondition.

While I was in there, I cleaned up the documentation for the square type.