add a command-line client
Because the web client doesn't show bindings and possible values for all squares at once, it's inconvenient for logic debugging (such as was needed in the fix of issue #32).  So I added a command line client that uses the pretty-printed string version of puzzles.  Then I realized that it would be handy to have a markdown version of a pretty-printed puzzle, for use in documentation.  So I added those and made the command line client understand them.  The markdown version of a puzzle includes row and column headers, but I haven't added those to the pretty-printed version (and the command line client doesn't yet allow them in assigns).