Provide utilities for data clear/re-initialize.
Provide a top-level command for cache and database clearing and
reloading.  This makes it possible to re-initialize from the Heroku
environment when there are database schema changes.  (Of course it can
also be run from the local environment with appropriate environment
variables, but it's more robust to be able to run it from a worker.)

This makes it unnecessary to reinitialize data on each deploy.  Those
deploys that need re-initialization can run this utility as part of the
deploy.