Clean up and fix the build.
So the Travis CI build started failing with what was clearly some form
of transactional inteference between various pieces of the test.  I
could repro this locally using their build script, so there was clearly
an issue that wasn't related to their environment (unlike the prior
postgres version issue).  I did a bunch of careful code review to see if
I was doing something stupid, which led to some clean up in the
sequencing and modularity of database and cache initialization.  But
ultimately the problem turned out to be Go's default use of parallel
processes to do testing on a multi-processor machine.  Adding -p 1 to
turn off this parallelism has resolved the problem locally, and I'm
hoping it will do so in Travis CI as well.