Introduce content hashes on puzzles and summaries.
As preparation for storing puzzles by their content signature, introduce
content hashes on puzzles and summaries.  I am using md5 since security
is not an issue, and it's cheap and small.  The hash is computed on the
geometry and the values of the puzzle, in case there can be two
different geometries over the same value array.  We convert the values
to bytes in order to compute the hash, so that limits us to side lengths
of 255, but that was already a limitation.