Update puzzle Signatures to be strings
Arrays of bytes are too tricky for clients to deal with; it's safer to
give them the hex strings of the signatures.  That simplifies database
and cache formats as well as use of signatures in URLs.  Plus, since
signatures are read-only, there's no need to convert them back to bytes.