fix #40 (improved puzzle display)
At first I thought the home page template could calculate the number of empty squares, so I remodeled the template loading process to allow for Funcs to be on templates.  But then it turned out that the required info wasn't actually making it onto the PuzzleInfo in the first place, so I had to go back and enhance PuzzleInfo to have it.