final touches: fix #41, fix #40
Really it's the earlier commits that fixed #41, but this commit is a
nice way to get the issue closed for release 1.0.

There were a lot of fixes for bugs in the fix of #40:
- improved the sorting to prioritize worked puzzles over unworked puzzles.
- enhance the web view to separate worked from unworked puzzles.

Also there was a fix introduced in the handling of the debug flag, which
had been defaulted to true when it should have been false.  And the
DEBUG env var was introduced for the same effect via 12-factor controls.